+++ lambdahub/App.js
@@ -1,20 +1,23 @@
-import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View } from 'react-native';
+import 'react-native-gesture-handler';
+import { NavigationContainer } from '@react-navigation/native';
+import { createDrawerNavigator } from '@react-navigation/drawer';
+import AnnouncementScreen from './screens/AnnouncementScreen';
+import CalendarScreen from './screens/CalendarScreen';
+import DriverScreen from './screens/DriverScreen';
+import HomeScreen from './screens/HomeScreen';
+
+
+const Drawer = createDrawerNavigator();
 
 export default function App() {
-  return (
-    <View style={styles.container}>
-      <Text>LambdaHub Home Page!-calendar</Text>
-      <StatusBar style="auto" />
-    </View>
-  );
+    return(
+        <NavigationContainer>
+            <Drawer.Navigator>
+                <Drawer.Screen name="Home" component={HomeScreen} />
+                <Drawer.Screen name="Announcements" component={AnnouncementScreen} />
+                <Drawer.Screen name="Calendar" component={CalendarScreen} />
+                <Drawer.Screen name="Driver Queue" component={DriverScreen} />
+            </Drawer.Navigator>
+        </NavigationContainer>
+    )
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
