+++ lambdahub/screens/HomeScreen.js
@@ -1,0 +1,18 @@
+import { View, Text, StyleSheet } from "react-native";
+
+export default function HomeScreen() {
+    return (
+      <View style={styles.container}>
+        <Text>LambdaHub Home Screen</Text>
+      </View>
+    );
+  }
+  
+  const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      backgroundColor: '#fff',
+      alignItems: 'center',
+      justifyContent: 'center',
+    },
+  });
